refactor: lifecycle mod
Plus additional code review feedback
